pyproject.toml
[build-system]
requires = [
  "setuptools>=61.0",
  "certifi>=14.05.14",
  "six>=1.10",
  "python_dateutil>=2.8.2",
  "urllib3>=2.1.0,<3.0.0",
  "prance>=23.6.21.0",
  "openapi-spec-validator>=0.7.1",
  "jsonschema>=4.23.0",
  "pyyaml>=5.2b1",
  "appdirs>=1.4.4",
  "pyjwt>=2.10.1",
  "pydantic>=2",
  "typing-extensions>=4.7.1",
]
build-backend = "setuptools.build_meta"

[project]
name = "recodex_cli_lib_eceltov"
version = "0.0.52"
authors = [
  { name="Vojtech Kloda", email="vojtech.kloda@gmail.com" },
]
description = "A library for ReCodEx API calls."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
  "setuptools>=61.0",
  "certifi>=14.05.14",
  "six>=1.10",
  "python_dateutil>=2.8.2",
  "urllib3>=2.1.0,<3.0.0",
  "prance>=23.6.21.0",
  "openapi-spec-validator>=0.7.1",
  "jsonschema>=4.23.0",
  "pyyaml>=5.2b1",
  "appdirs>=1.4.4",
  "pyjwt>=2.10.1",
  "pydantic>=2",
  "typing-extensions>=4.7.1",
]

[project.urls]
Homepage = "https://github.com/ReCodEx"
Issues = "https://github.com/ReCodEx"
